Move Dense matrix formatting functions to a separate file
